Rework Inductor support for collectives.
This is done by introducing two new base classes: InPlaceCollectiveKernel and OutOfPlaceCollectiveKernel.

They deal with the differences for when InPlaceHint needs to be used.

Additionally to that, we introduce `has_side_effects` method to buffers that
prevents them from being DCE'd by the scheduduler. This is needed because InPlaceHint
nodes both wrap the inputs and are the outputs, which places no users to the collectives
themselves.

ghstack-source-id: e5d132370a27c75908db6719376c35274ded4d14
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99765